requirements.txt
streamlit
langchain_groq
langchain-community
langchain-core
faiss-cpu
python-dotenv
PyPDF2
pypdf
langchain-objectbox
langchain